Refactor DXF group code value handling for type safety
Modernize EoDxfGroupCodeValuesVariant to support all DXF group code value types using std::variant and type-safe accessors. Update DXF reader and writer classes for explicit type handling, ensuring correct parsing and writing of 16-bit, 32-bit, 64-bit, boolean, and other types. Refactor header, entity, and table parsing to use new type-safe methods. Remove obsolete code, improve documentation, and lay groundwork for future extensibility while maintaining backward compatibility.

+++ Core Libraries/EoDxfLib/EoDxfObjects.cpp
@@ -109,10 +109,10 @@ void EoDxfImageDefinition::ParseCode(int code, EoDxfReader* reader) {
       m_vSizeOfOnePixel = reader->GetDouble();
       break;
     case 280:
-      m_imageIsLoadedFlag = reader->GetInt32();
+      m_imageIsLoadedFlag = reader->GetInt16();
       break;
     case 281:
-      m_resolutionUnits = reader->GetInt32();
+      m_resolutionUnits = reader->GetInt16();
       break;
     default:
       EoDxfObjectEntry::ParseCode(code, reader);
